Calculate average evolutionary distinctiveness across multiple time scales

This function calculates the average evolutionary distinctiveness for a list of species across distinct time slices.

Usage

avg_evol_dist(
  phy,
  data_frame,
  taxon = NULL,
  time_slices = c(25, 50, 75, 100, 115),
  num_cores = 1
)

Arguments

phy

A rooted, time-calibrated (branch lengths representative of time) phylogenetic tree used to calculate evolutionary distinctiveness.

data_frame

A data frame containing species data with a "species_name" column.

taxon

A character vector specifying the taxa (species) of interest. Default is NULL.

time_slices

A numeric vector specifying the desired time slices to consider. Default is 25 MYA, 50 MYA, 75 MYA, 100MYA, 115MYA. These should be adjusted based on the time-extent of the phylogeny; the default times are based on a large phylogeny of Angiosperms. If the chosen time slices are not reflected on the utilized phylogeny, an error will occur.

num_cores

Number of CPU cores to use for parallel processing. Default is 1. We recommend utilizing as many cores as possible.

Value

A data frame with species names and their corresponding evolutionary distinctiveness values. Species not included in the phylogeny will return NA.

Details

Many species are evolutionarily distinct within genera but are evolutionarily indistinct across larger phylogenetic extents. This function accounts for differences in evolutionary distinctiveness at different time extents through the calculation of evolutionary distinctiveness at five distinct phylogenetic extents.

Examples