Build Trait Data from the LEDA Traitbase

Build a dataframe of trait data for various plant species of interest using LEDA trait database.

Usage

build_trait_data_LEDA(columns_to_select, genera = NULL)

Arguments

columns_to_select

A character vector specifying which traits to select. Options include: SLA (specific leaf area mm^2/mg), seed_mass (mg), leaf_mass (mg), and canopy_height (m)

genera

A character vector specifying genera of interest to be filtered. Default is NULL.

Value

A dataframe including mean trait data for all species or species of choice in respective columns titled by trait name.

Details

The function reads trait data from four different data files, including SLA (Specific Leaf Area), seed mass, leaf mass, and canopy height. It then combines these data frames into a single data frame based on the species name.

References

kleyer2008GeoFunPhy

Author

Alivia G Nytko, anytko@vols.utk.edu

Examples