Calculate range sizes for multiple species

This function calculates the range sizes (sq km) of multiple species by retrieving occurrence data from GBIF, generating convex hulls, and clipping them to continent boundaries. It uses parallel processing to handle multiple species efficiently.

Usage

calc_range_size(
  data_frame,
  num_cores = 1,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000,
  continent_file = NULL
)

Arguments

data_frame

A data.frame containing scienfic names of species with a column named species_name.

num_cores

The number of cores to use for parallel processing. Default is 1 (no parallel processing).

min_points

The minimum number of points required to form a range cluster. Default is 5.

min_distance

The minimum distance between points in a range cluster. Default is 1 epsilon (eps).

gbif_limit

The maximum number of GBIF records to retrieve per species. Default is 2000.

continent_file

The URL of a GeoJSON or shape (.shp & .shx) file containing country or continent data. Default is NULL - If NULL, a GeoJSON file from Natural Earth will be used.

Value

A data.frame with two columns: species_name and range_size. The range_size column represents the size of the range for each species in square kilometers.

Examples

# Generate a test dataframe with species Abies cephalonica
test_data <- data.frame(species_name = c("Betula_humilis"))

# Generate range sizes for each speices
range_size <- calc_range_size(data_frame = test_data, min_points = 4)
#> Processing species:Betula_humilis
print(range_size)
#>     species_name range_size
#> 1 Betula_humilis   170668.2