Analyze Species Convex Hulls Against Biome Data

This function checks which biomes are intersected by the convex hulls of species distributions. By default, the function uses a preloaded biome shapefile containing four biomes: Humid Tropics, Dry Tropics, Temperate, and Boreal. Users can also provide their own biome data as an sf object.

Usage

check_biomes(convex_hulls, biome_sf = NULL)

Arguments

convex_hulls

A named list of convex hulls for each species. Each element of the list should contain a list of sf polygon objects representing the convex hulls of species' distributions.

biome_sf

An optional sf object representing the biome polygons. If not provided, the default biome shapefile (biomes.shp) from the package will be used. The default biomes are:

Humid Tropics

Dry Tropics

Temperate

Boreal

Value

A dataframe where each row corresponds to a species, and each column (except species_name) represents a biome. A value of 1 in the cell indicates that the species' convex hull intersects with the respective biome.

Note

If the CRS of the provided biome_sf or convex_hulls is not set, the function will automatically assign and transform them to EPSG:4326 (WGS84).

Examples