Clip species range polygons to land boundaries

This function uses a continent GeoJSON or shape file (.shp & .shx) file to clip range polygons to land boundaries.

Usage

clip_polygons_to_land(convex_hulls, continent_sf)

Arguments

convex_hulls

A nested list of sf objects representing the convex hulls of a species range.

continent_sf

A file of continent boundaries. We recommend using Natural Earth's GeoJSON or shape files.

Value

A nested list of clipped range polygons

Details

When clipping the polygons of species with no points touching land (i.e. an oceanic polygon) the function will return an error. This error has been silenced and the oceanic polygon will be removed from the clipped polygon list.

Examples