Compute Within-Cluster Sum of Squares (WCSS) Compute the Within-Cluster Sum of Squares (WCSS) for a given variable and a range of cluster values (k) using k-means clustering.

Compute Within-Cluster Sum of Squares (WCSS) Compute the Within-Cluster Sum of Squares (WCSS) for a given variable and a range of cluster values (k) using k-means clustering.

Usage

compute_wcss(data, variable, k_values, seed = 42)

Arguments

data

A data frame containing the data for clustering.

variable

The variable in the data frame to be used for clustering. Use the name of the column in "".

k_values

A vector of integers specifying the number of clusters (k) to be considered.

Value

A numeric vector containing the WCSS values for each value of k.

Details

This function computes the Within-Cluster Sum of Squares (WCSS) for each value of k provided in the k_values parameter using k-means clustering. It returns a numeric vector where each element represents the WCSS value for the corresponding value of k.

See also

Examples