Get continent spatial data from a GeoJSON file

This function reads a GeoJSON file from a given URL, extracts continent information, and returns an sf object with continent polygons.

Usage

get_continent_sf(url = NULL)

Arguments

url

Optional URL of a GeoJSON file containing country or continent data. Default is NULL - if NULL the package uses embeded continent boundaries from Natural Earth.

Value

An sf object containing polygons representing continents. The object includes a column for the continent names and corresponding geometries.

Examples

# Get continent data from a GeoJSON file
continent_sf <- get_continent_sf()
print(continent_sf)
#> Simple feature collection with 8 features and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -89.99893 xmax: 180 ymax: 83.59961
#> Geodetic CRS:  WGS 84
#>                 continent                       geometry
#> 1                  Africa MULTIPOLYGON (((-3.025879 5...
#> 2              Antarctica MULTIPOLYGON (((-163.6018 -...
#> 3                    Asia MULTIPOLYGON (((54.12949 12...
#> 4                  Europe MULTIPOLYGON (((-52.00293 4...
#> 5           North America MULTIPOLYGON (((-155.9884 1...
#> 6                 Oceania MULTIPOLYGON (((-176.2208 -...
#> 7 Seven seas (open ocean) MULTIPOLYGON (((-26.45103 -...
#> 8           South America MULTIPOLYGON (((-74.66597 -...