Get Convex Hull Polygons for Species from GBIF Data

This function gathers, cleans, and clusters GBIF data to create convex hulls of species ranges.

Usage

get_range_convex_hulls(
  data_frame,
  species_name = NULL,
  num_cores = 1,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000
)

Arguments

data_frame

A dataframe containing scientific names of species. Must include a column named species_name.

species_name

An optional character vector of species names to process. Default is NULL. If NULL, all species in the dataframe are processed.

num_cores

The number of cores to use for parallel processing. Default is 1 (no parallel processing).

min_points

The minimum number of points required to form a range cluster. Default is 5.

min_distance

The minimum distance between points in a range cluster. Default is 1 epsilon (eps).

gbif_limit

The maximum number of GBIF records to retrieve per species. Default is 2000.

Value

A nested list of convex hull polygons for each species. Each species name corresponds to a list of sf objects representing the convex hulls.

Examples

# Generate a test dataframe with species Abies cephalonica
test_data <- data.frame(species_name = c("Abies_cephalonica"))
# Retrieve the unclipped range polygons 
unclipped_hulls <- get_range_convex_hulls(test_data)
#> Processing species:Abies_cephalonica
print(unclipped_hulls)
#> $Abies_cephalonica
#> $Abies_cephalonica$`1`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 22.17968 ymin: 36.94886 xmax: 23.90966 ymax: 38.76327
#> Geodetic CRS:  WGS 84
#> POLYGON ((22.90655 36.94886, 22.98478 36.95195,...
#> 
#> $Abies_cephalonica$`2`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 20.57506 ymin: 38.13723 xmax: 20.67796 ymax: 38.29078
#> Geodetic CRS:  WGS 84
#> POLYGON ((20.67278 38.13723, 20.67796 38.15113,...
#> 
#> $Abies_cephalonica$`3`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 14.23305 ymin: 56.12064 xmax: 14.23323 ymax: 56.12079
#> Geodetic CRS:  WGS 84
#> POLYGON ((14.23305 56.12065, 14.23306 56.12064,...
#> 
#> $Abies_cephalonica$`4`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0.2926 ymin: 46.62728 xmax: 0.94663 ymax: 47.43229
#> Geodetic CRS:  WGS 84
#> POLYGON ((0.2926 46.86807, 0.41926 46.62728, 0....
#> 
#> $Abies_cephalonica$`5`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 1.99685 ymin: 42.39377 xmax: 7.40942 ymax: 47.60943
#> Geodetic CRS:  WGS 84
#> POLYGON ((1.99685 42.44291, 2.46658 42.39377, 6...
#> 
#> $Abies_cephalonica$`6`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 1.20039 ymin: 44.83076 xmax: 1.96516 ymax: 45.53002
#> Geodetic CRS:  WGS 84
#> POLYGON ((1.20039 44.88326, 1.35315 44.83076, 1...
#> 
#> $Abies_cephalonica$`7`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -1.07394 ymin: 42.86369 xmax: -0.59664 ymax: 42.9991
#> Geodetic CRS:  WGS 84
#> POLYGON ((-0.59664 42.86369, -0.84351 42.98312,...
#> 
#> $Abies_cephalonica$`8`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2.3259 ymin: 47.55198 xmax: 2.68092 ymax: 48.84401
#> Geodetic CRS:  WGS 84
#> POLYGON ((2.35995 48.84401, 2.3259 47.55198, 2....
#> 
#> $Abies_cephalonica$`9`
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 6.34018 ymin: 48.39447 xmax: 7.4431 ymax: 48.7513
#> Geodetic CRS:  WGS 84
#> POLYGON ((6.34018 48.7513, 6.80616 48.40738, 7....
#> 
#>