Create Interactive 3d Scatter Plot

Create an interactive scatter plot depicting the ecological status of each species in 3d space

Usage

plot_EER_status(data_frame, fun_dist, evol_dist, range_size)

Arguments

data_frame

A data frame containing species data.

fun_dist

A column within the data frame corresponding to the functional distinctivess of the species

evol_dist

A column within the data frame corresponding to the evolutionary distinctivess of the species; should be z transformed (typically using the median and meadian absolute deviation) for ecological status classification.

range_size

A column within the data frame corresponding to the range size (km^2) of the species; should be z transformed (typically using the median and meadian absolute deviation) for ecological status classification.

Value

An interactive 3d rgl object that visualizes the three dimensions of rarity and resultant ecological status. To capture a still image of the scatter plot rgl::rgl.snapshot() is recommended.

Details

The function checks if the specified fun_dist, evol_dist, and range_size columns exists in the data frame and if the required columns 'classifications,' and 'species_name' are present. It then assigns colors to data points based on classifications and displays the 3D scatter plot interactively.

See also

Author

Alivia G Nytko, anytko@vols.utk.edu

Examples

# Create dataframe including species names, range size, evolutionary distinctiveness, functional distinctiveness, and their respective eco-evolutionary classifications

species_names <- c("Acer_campestre", "Acer_monspessulanum", "Acer_negundo", "Acer_opalus", "Acer_platanoides", "Acer_pseudoplatanus", "Acer_saccharinum")
FD_values <- runif(min = -2, max = 2, n=7)
range_values <- runif(min = -2, max = 2, n=7)
evol_dist_values <- runif(min = -2, max = 2, n=7)
classification <- c("Common", "Common", "Endemic", "Keystone", "Competitively Rare", "Common", "Adaptable Survivor")

maple_data_figure <- data.frame(species_name = species_names, fun_dist = FD_values, range_size = range_values, evol_dist = evol_dist_values, classifications = classification)

# Plot the species' classifications in 3D space
interactive_plot <- plot_EER_status(data_frame = maple_data_figure, fun_dist = "fun_dist", range_size = "range_size", evol_dist = "evol_dist")