Plot clipped convex hulls on map

Plot clipped convex hulls on map

Usage

plot_clipped_hulls(clipped_convex_hulls)

Arguments

clipped_convex_hulls

A list of clipped convex hulls for each species.

Details

This function plots a list of clipped convex hulls onto a world leaflet map.

Examples

data <- data.frame(species_name = c("Eucalyptus gunnii", "Eucalyptus tenuiramis", "Eucalyptus globulus"))
continent_bounds <- get_continent_sf()
euc_convex_hulls <- get_range_convex_hulls(data)
#> Processing species:Eucalyptus gunnii
#> Processing species:Eucalyptus tenuiramis
#> Processing species:Eucalyptus globulus
euc_clipped_hulls <- clip_polygons_to_land(euc_convex_hulls, continent_bounds)
plot_clipped_hulls(clipped_convex_hulls = euc_clipped_hulls)