Plot range polygons on world map

Plot range polygons on world map

Usage

range_poly_map(
  data_frame,
  species_name = NULL,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000,
  plot = TRUE,
  clip = FALSE,
  continent_sf = NULL
)

Arguments

species_name

A character vector of species names.

min_points

The minimum number of points required to form a range cluster. Default is 5.

min_distance

The minimum distance between points in a range cluster. Default is 1 epsilon (eps).

gbif_limit

The maximum number of GBIF records to retrieve per species. Default is 2000.

plot

Logical indicating whether to plot the convex hulls on a map.

Value

A leaflet map of species range polygons.

Details

This function combines the functionality of EcoStatusR::create_range_polygons() and EcoStatusR::plot_convex_hulls(). Specifcally this function creates convex hulls for species using occurence data from GBIF and plots them onto a world leaflet map.

Examples

species_data <- data.frame(species_name = c("Abies alba", "Quercus nigra"))

# Plotting unclipped range polygons
range_poly_map(data_frame = species_data)
#> Processing species:Abies alba
#> Processing species:Quercus nigra
#> Warning: minimal value for n is 3, returning requested palette with 3 different levels

# Plotting clipped range polygons
continent_bounds <- get_continent_sf()
range_poly_map(data_frame = species_data, clip = TRUE, continent_sf = continent_bounds)
#> Processing species:Abies alba
#> Processing species:Quercus nigra
#> Warning: minimal value for n is 3, returning requested palette with 3 different levels