Scale by Median

Transform values in a dataframe into z scores around the median. To classify species status, measures of range size and evolutionary distinctiveness must be tranformed.

Usage

scale_by_median(data_frame, columns_chosen = NULL)

Arguments

data_frame

A data frame containing the columns to transform.

columns_chosen

A character vector specifying the column names to transform. Default is NULL.

Value

A scaled data frame with Z-scores computed using the median and medain absolute deviation (MAD) for the selected columns.

Details

Chosen columns are scaled using the median and median absolute deviation. This transformation is recommend for range size and evolultionary distinctiveness when classifying the ecological status of species using EcoStatusR::check_eco_status().

See also

Author

Alivia G Nytko, anytko@vols.utk.edu

Examples